Move single threaded builder struct
